+++ Cargo.toml
@@ -11,14 +11,26 @@ name = "datagen"
 path = "src/bin/datagen.rs"
 required-features = ["datagen"]
 
+[[bin]]
+name = "wiz"
+path = "src/bin/wiz/main.rs"
+required-features = ["wiz"]
+
 [features]
-datagen = ["serde", "serde_json", "ctrlc", "num_cpus"]
+datagen = [
+  "clap",
+  "ctrlc",
+  "num_cpus",
+  "serde",
+  "serde_json",
+]
+wiz = ["clap"]
 
 [dependencies]
 anyhow = "1.0.86"
 arrayvec = "0.7.4"
 bitflags = "2.6.0"
-clap = { version = "4.5.15", features = ["derive"] }
+clap = { version = "4.5.15", features = ["derive"], optional = true }
 ctrlc = { version = "3.4.5", features = ["termination"], optional = true }
 num_cpus = {version =  "1.16.0", optional = true }
 rand = { version = "0.9.0", features = ["small_rng"] }
