+++ src/chess/position/fen.rs
@@ -222,7 +222,7 @@ impl CastleRights {
 }
 
 #[cfg(test)]
-mod test {
+mod tests {
     use super::*;
 
     #[test]
