+++ src/engine.rs
@@ -6,4 +6,9 @@ pub mod tt;
 pub mod uci;
 
 mod movepicker;
+mod search_manager;
+mod time_management;
 mod utils;
+
+pub use search_manager::SearchManager;
+pub use uci::Uci;
