+++ src/engine/movepicker.rs
@@ -212,7 +212,7 @@ impl MovePicker {
 }
 
 #[cfg(test)]
-mod test {
+mod tests {
     use crate::chess::movegen::init_tables;
     use crate::chess::position::fen::Fen;
     use crate::zobrist::init_zobrist;
